Update math.js too

+++ math.js
@@ -4,5 +4,5 @@ exports.isStar = true;
 
 exports.sum = function (a, b) {
     // Реализуйте сложение в этой функции
-    return parseInt(a) + parseInt(b);
+    return parseFloat(a) + parseFloat(b);
 };
